Introduce response filters that are able to process HTTP/1.1 responses. This introduces several 'base' filters:
- HTTP Status code filter
- HTTP header filter
- HTTP (de)chunking filter

A rough, non-completed implementation of deflate filters are also present.
(I know what would need to be done, but the code is a bit much right now
for that.)

Some rework has also been done to serf_get to add these new filters and
streamline the code a little bit.  (Note the #defines which control its
verbosity.)

Since Serf is now conditionally HTTP/1.1-compliant, bump the version up to
0.0.2-dev.


git-svn-id: https://svn.apache.org/repos/asf/commons/serf/trunk@30 13f79535-47bb-0310-9956-ffa450edef68