Fix https onnection abort issue when using https. This should fix Subversion issue 4174: http://subversion.tigris.org/issues/show_bug.cgi?id=4174
When the server sends us a "close notify" SSL Alert indicating that it's 
going to close the socket, we have to stop sending on that socket immediately.

Failing to do so will make the server reset the connection, which can result
in an ECONNABORTED error upon next read or write. This seems to happen esp.
on Windows.

* buckets/ssl_buckets.c
  (ssl_decrypt): when SSL_read returns ok but with length 0, a "close notify"
   alert was received. Return APR_EOF to let upstream know that we're done
   with this socket.

* outgoing.c
  (read_from_connection): Add comment to note that SSL alert's are also an
   example of unexpected incoming data.


git-svn-id: https://svn.apache.org/repos/asf/serf/trunk@1699165 13f79535-47bb-0310-9956-ffa450edef68