serf_bucket_headers_get() is inherently broken: the caller doesn't know whether it must call mem_free() on the result or not.
Thankfully, this really only makes a different when fetching
*-Authenticate headers. And even then... most allocators are never
queried if unfreed items exist. And... the memory gets tossed with the
pool. But: this is still an API bug.

For now: move the *-Authenticate query into a debug space. Leave a
marker in the _get() function. Revisit in the future.

* buckets/headers_buckets.c:
  (serf_bucket_headers_get): leave a marker that this is broken.

* auth/auth.c:
  (dispatch_auth): move the logging into a block enclosed by
    AUTH_VERBOSE. just rely on the gather-into-hash to see if there
    are headers. (we don't have to fast-path check for headers since
    they *should* be there, per spec)


git-svn-id: https://svn.apache.org/repos/asf/serf/trunk@1699638 13f79535-47bb-0310-9956-ffa450edef68