Fix issue #120: Error with ssl tunnel over proxy with KeepAlive off and Basic authentication.
Don't requeue the CONNECT request on a new connection after the old connection
was reset by an 407 response. A CONNECT request will be automatically created
when setting up a new connection, if the queue already had one from a previous
connection, that extra request will be sent to the server over the ssl tunnel.
(which is wrong).

This also fixes an (not reported) issue where the application's credentials
callback function is called with requeued CONNECT request (which the app doesn't
know) an a NULL baton, potentially causing crashes in the application.

* test/test_context.c
  (ssltunnel_basic_auth): Take a new parameter proxy_resp_hdrs and add it
       value as headers to the responses of the proxy.
  (test_ssltunnel_basic_auth,
   test_ssltunnel_basic_auth_server_has_keepalive_off): Update call to 
       ssltunnel_basic_auth with an empty proxy header.
  (test_ssltunnel_basic_auth_server_has_keepalive_off): New test.
  (test_context): Run the new test.

* outgoing.c
  (reset_connection): When requeueing requests on a new connection, filter
       out the CONNECT request.


git-svn-id: https://svn.apache.org/repos/asf/serf/trunk@1699659 13f79535-47bb-0310-9956-ffa450edef68